+++ __windows__/_common_/js/parag.js
@@ -35,10 +35,24 @@ class Parag
         updated_at: moment().format()
       , last_parag_id: this._lastID
     })
-    console.log('this._lastID',this._lastID)
+    // console.log('this._lastID',this._lastID)
     return Number(this._lastID)
   }
 
+  /**
+  * Méthode pour tout déselectionner (obtenue principalement en cliquant
+  * en dehors des paragraphes).
+  * Cela supprimer aussi l'élément courant
+  **/
+  static deselectAll ()
+  {
+    if ( this.current ) { this.unsetCurrent() }
+    if ( this._selecteds )
+    {
+      this._selecteds.forEach( iparag => iparag.setDeselected.bind(iparag)() )
+      this._selecteds = []
+    }
+  }
 
   // Met le paragraphe +iparag+ en paragraphe courant
   static setCurrent (iparag)
@@ -147,6 +161,47 @@ class Parag
     return this.data
   }
 
+  /**
+  * @return {Array} de {Parag}, la liste des relatifs du parag courant
+  *         tels que définis dans les données Relatives du projet
+  **/
+  get relatifs ()
+  {
+    if ( undefined === this._relatifs )
+    {
+      let drelatifs = this.data_relatives
+      let arr = {as_referent: [], as_relatifs: []}
+        , pan_arr
+        , is_referent
+
+      for ( let pan in drelatifs['r'] )
+      {
+        pan_arr = drelatifs['r'][pan]
+        // On détermine si le parag courant est référent en déterminant le nombre
+        // d'élément dans sa donnée panneau.
+        // Par exemple, si la donnée "n" (notes) contient 2 identifiants, alors
+        // on considère que le parag courant est référent et que ces deux
+        // notes sont les relatifs purs
+        is_referent = pan_arr.length > 1
+        // On enregistre les relatifs dans la liste as_relatifs si le parag
+        // est référent est dans la liste as_referent dans le cas contraire
+        if ( is_referent )
+        {
+          pan_arr.forEach( pid => arr.as_relatifs.push(Parags.get(pid)))
+        }
+        else
+        {
+          arr.as_referent.push(Parags.get(pan_arr[0]))
+        }
+      }
+      this._relatifs = arr
+    }
+    return this._relatifs
+  }
+
+  get data_relatives ()
+    { return Projet.current.relatives.data.relatives[String(this.id)] }
+
   edit ()
   {
     Parag.setCurrent(this)
@@ -238,9 +293,32 @@ class Parag
   deselect ()
   {
     if ( ! this.selected ) { return this }
-    DOM.removeClass(this.mainDiv, 'selected')
     Parag.supSelect(this)
+    return this.setDeselected()
+  }
+  // Méthode de déselection qui peut être appelée quane on doit
+  // tout déselectionner, pour ne pas utiliser la méthode Parag.supSelect
+  setDeselected ()
+  {
+    DOM.removeClass(this.mainDiv, 'selected')
     this.selected = false
+    if ( this.relatifsExergued ) { this.unexergueRelatifs() }
+    return this
+  }
+
+  exergue (as_referent)
+  {
+    if ( this.exergued ) { return this }
+    DOM.addClass(this.mainDiv, as_referent ? 'referent' : 'relatif')
+    this.exergued = true
+    return this
+  }
+  unexergue ()
+  {
+    if ( ! this.exergued ) { return this }
+    DOM.removeClass(this.mainDiv, 'referent')
+    DOM.removeClass(this.mainDiv, 'relatif')
+    this.exergued = false
     return this
   }
 
@@ -325,12 +403,18 @@ class Parag
   {
     if (evt.metaKey)
     {
-      console.log('Click avec la touche CMD')
+      // CMD Click
       if ( Projet.mode_double_panneaux )
       {
-        console.log('Click avec la touche CMD en mode double panneaux')
-        if ( this.selected ){ this.deselect() }
-        else { this.select() }
+        // <= CMD Click mode double panneaux
+        if ( this.selected ){
+          this.deselect()
+          this.unexergueRelatifs()
+        }
+        else {
+          this.select()
+          this.exergueRelatifs()
+        }
       }
       return DOM.stopEvent(evt)
     }
@@ -338,6 +422,100 @@ class Parag
 
   // --- private ---
 
+  /**
+  * Méthode qui met en exergue tous les relatifs du paragraphe
+  * Plus précisément, la méthode va regarder si le paragraphe est un relatif
+  * ou un référent et mettre les paragraphes dans le style en fonction.
+  * Le paragraphe est un référent s'il est seul dans la données de ses
+  * relatif et il est relatif s'il n'est pas seul dans cette donnée.
+  **/
+  exergueRelatifs ()
+  {
+    this.relatifs.as_relatifs.forEach( iparag => iparag.exergue(false) )
+    this.relatifs.as_referent.forEach( iparag => iparag.exergue(true) )
+    // Il faut régler la classe du paragraphe cliqué en fonction
+    // de son statut dans les doubles panneaux, referent ou relatif
+    let statutDblPan = this.statutDoublePanneau
+    if(statutDblPan !== 'none')
+    {
+      // Si le parag courant cliqué est un relatif, alors il faut mettre en
+      // exergue tous les relatifs du référent pour mettre en exergue les
+      // frère de ce parag.
+      // Noter que la propriété `this.referent_id` est tout à fait provisoire
+      // et dépend entièrement des panneaux qui sont activés.
+      if ( statutDblPan == 'relatif' )
+      {
+        Parags.get(this.referent_id).exergueRelatifs()
+      }
+      else
+      {
+        DOM.addClass(this.mainDiv, statutDblPan)
+      }
+    }
+    this.relatifsExergued = true
+  }
+  /**
+  * Méthode qui sert de l'exergue tous les relatifs du paragraphe
+  **/
+  unexergueRelatifs ()
+  {
+    this.relatifs.as_relatifs.forEach( iparag => iparag.unexergue() )
+    this.relatifs.as_referent.forEach( iparag => iparag.unexergue() )
+    let statutDblPan = this.statutDoublePanneau
+    if ( statutDblPan != 'none' )
+    {
+      if ( statutDblPan == 'relatif'){
+        Parags.get(this.referent_id).unexergueRelatifs()
+      }
+      else
+      {
+        DOM.removeClass(this.mainDiv, statutDblPan)        
+      }
+    }
+    this.relatifsExergued = false
+  }
+
+  /**
+  * Propriété statut du parag dans le mode double panneau actuel
+  * En fonction des deux panneaux choisis, le paragraphe peut avoir trois
+  * status différents :
+  *     'none'      Il n'a pas de relatives
+  *     'referent'  Il est le référent de plusieurs paragraphes de l'autre
+  *                 panneau.
+  *     'relatif'   Il est relatif, avec d'autres paragraphes à côté de
+  *                 lui, d'un référent de l'autre panneau.
+  **/
+  get statutDoublePanneau ()
+  {
+    if ( ! Projet.mode_double_panneaux ) { return 'none' }
+    let pan_letter = Projet.PANNEAUX_DATA[this.panneau_id].oneLetter
+    // console.log(`pan_letter du parag ${this.id} : ${pan_letter}`)
+    // Il faut récupérer les deux panneaux activés
+    let other_panneau
+    if ( Projet.alt_panneau.name == this.panneau_id ){
+      other_panneau = Projet.current_panneau.name
+    } else {
+      other_panneau = Projet.alt_panneau.name
+    }
+    // console.log(`Nom de l'autre panneau : ${other_panneau}`)
+    let other_pan_letter = Projet.PANNEAUX_DATA[other_panneau].oneLetter
+    // console.log(`Lettre de l'autre panneau : ${other_pan_letter}`)
+    let drelatives = this.data_relatives
+    // console.log(`Données relative : ${JSON.stringify(drelatives)}`)
+    if (undefined === drelatives['r'][other_pan_letter]) {return 'none'}
+    // console.log(`drelatives['r']['${other_pan_letter}'] = ${drelatives['r'][other_pan_letter].join(', ')}`)
+    if ( drelatives['r'][other_pan_letter].length > 1){
+      // console.log("=> referent")
+      return 'referent' // ce paragraphe est le référent
+    } else {
+      // Ce paragraphe est un relatif
+      // On va plutôt renvoyer le travail d'afficher les relatifs à son
+      // référent pour que tous les relatifs soient bien affichés.
+      this.referent_id = drelatives['r'][other_pan_letter][0]
+      // console.log("=> relatif")
+      return 'relatif'
+    }
+  }
 
   // Passer le champ contents en mode édition (sauf si la touche CMD est
   // pressée)