+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+# 0.1.6
+
+  * Sélection et déselection (mise en exergue) des éléments relatifs, avec une exergue particulière pour le référent. On doit être en mode double panneau pour atteindre cette fonctionnalité, et le référent est entouré de rouge tandis que les relatifs sont entourés de pourpre.
+
+# 0.1.5
+
+  * Un clic en dehors des paragraphes déselectionne tous les paragraphes sélectionnés.
+
 # 0.1.4
 
   * Modification du fonctionnement des `Relatives`.
-  
+
 # 0.1.3
 
   * Implémentation des `relatives` qui permettent de tenir à jour la relation entre les paragraphes de deux `PanProjets` différents, c'est-à-dire entre par exemple le synopsis et le manuscrit.